calculate sinceDateTime based on last run
use java.time.* instead of java.util.Date

+++ smartclide-monitoring/src/main/java/de/atb/context/monitoring/models/GitlabCommitMessage.java
@@ -37,5 +37,5 @@ public class GitlabCommitMessage {
     String repository;
     String branch;
     Integer noOfModifiedFiles;
-    Integer timeSinceLastCommit;
+    Long timeSinceLastCommit;
 }
